# part: Improve node4 find speed by ~2%