Change the loop into a switch. We don't need to check the size as all operations make sure to nil out the child.

Before:
BenchmarkDB_RandomLookup-6                         27901             42534 ns/op          23510630 objects/sec
BenchmarkDB_SequentialLookup-6                     35810             33500 ns/op          29851094 objects/sec

After:
BenchmarkDB_RandomLookup-6                         28411             41583 ns/op          24048128 objects/sec
BenchmarkDB_SequentialLookup-6                     36318             32940 ns/op          30358678 objects/sec

41583/42534 = 0.977
32940/32940 = 0.98i3